smart_agent/requirements.txt
# fastapi_session==0.1.1
PyYAML
Requests==2.31.0
boto3
botocore
fastapi
mangum
openai
pydantic==1.10.10
python-dotenv==1.0.0
requests
uvicorn
validators==0.20.0
